requirements.txt
asgiref==3.4.1
boto3==1.21.9
botocore==1.24.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Django==4.0.2
django-cors-headers==3.10.1
django-redis==5.1.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
jmespath==0.10.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pdf2image==1.16.0
Pillow==9.1.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.8
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
whitenoise==5.3.0
